refactor: inner class Position public class로 변환
