+++ src/main/java/racingcar/car/Position.java
@@ -1,0 +1,22 @@
+package racingcar.car;
+
+public class Position {
+
+    private int position;
+
+    private Position(int position) {
+        this.position = position;
+    }
+
+    public static Position from(int position){
+        return new Position(position);
+    }
+
+    public void move(){
+        this.position++;
+    }
+
+    public int getPosition(){
+        return position;
+    }
+}
